resources/test/tests/apisample15.html

<!DOCTYPE HTML>
<html>
<head>
<title>Example with a shared worker</title>
<script src="../../testharness.js"></script>
<script src="../../testharnessreport.js"></script>
</head>
<body>
<h1>Shared Web Worker Tests</h1>
<p>Demonstrates running <tt>testharness</tt> based tests inside a shared worker.
<p>The test harness should time out due to one of the tests in the worker timing out.
<p>This test assumes that the browser supports <a href="http://www.w3.org/TR/workers/#shared-workers-and-the-sharedworker-interface">shared web workers</a>.
<div id="log"></div>

<script>
test(
    function(t) {
        assert_true("SharedWorker" in self,
                    "Browser should support SharedWorkers");
    },
    "Browser supports SharedWorkers");

fetch_tests_from_worker(new SharedWorker("apisample-worker.js",
                                         "My shared worker"));
</script>
<script type="text/json" id="expected">
{
  "status": {
    "status": 2,
    "OK": 0,
    "TIMEOUT": 2,
    "ERROR": 1,
    "message": null,
    "stack": null
  },
  "tests": [
    {
      "status": 0,
      "index": 1,
      "name": "Browser supports SharedWorkers",
      "properties": {},
      "TIMEOUT": 2,
      "PASS": 0,
      "FAIL": 1,
      "message": null,
      "NOTRUN": 3,
      "stack": null
    }
  ],
  "type": "complete"
}
</script>
</body>
